scripts/requirements.txt
alembic==1.11.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector==2.2.9
python-dotenv==1.0.0
SQLAlchemy==2.0.15
typing-extensions==4.6.3
Werkzeug==2.3.4
WTForms==3.0.1
zipp==3.16.2
